third_party/blink/web_tests/external/wpt/client-hints/viewport-width-window.https.html

<html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(t => {
  return new Promise(resolve => {
    window.addEventListener("message", t.step_func(e => {
      assert_equals(e.data.viewport, window.innerWidth.toString());
      resolve();
    }));
  });
});
window.open("resources/viewport-frame.py", "");
</script>
