third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/viewport-width-window.https-expected.txt

This is a testharness.js-based test.
FAIL viewport-width-window assert_equals: expected "800" but got "FAIL"
Harness: the test ran to completion.
